#3 added mappingEntityRule completion provider
